refactor: split processes into focused modules
Closes #129
