+++ src/main/processes/tasklist.ts
@@ -1,0 +1,22 @@
+import { execFile } from 'child_process'
+
+export function readRunningProcessNames() {
+  return new Promise<Set<string>>((resolve) => {
+    execFile('tasklist', ['/fo', 'csv', '/nh'], { windowsHide: true }, (error, stdout) => {
+      if (error) {
+        console.error('Failed to read running processes:', error)
+        resolve(new Set())
+        return
+      }
+
+      const names = new Set<string>()
+      stdout.split(/\r?\n/).forEach((line) => {
+        const match = line.match(/^"([^"]+)"/)
+        if (match) {
+          names.add(match[1].toLowerCase())
+        }
+      })
+      resolve(names)
+    })
+  })
+}
